Attribute data type change

+++ pojo/src/main/java/com/stanzaliving/core/leadership/dto/LeadershipPropertyAttributesDto.java
@@ -32,7 +32,7 @@ public class LeadershipPropertyAttributesDto {
 
 	private Integer rooms;
 
-	private Integer rentPerBed;
+	private long rentPerBed;
 
 	private Date llHandoverDate;
